NaN double

+++ src/main/java/acme/features/manager/dashboard/ManagerDashboardRepository.java
@@ -42,10 +42,10 @@ public interface ManagerDashboardRepository extends AbstractRepository {
 	double devUserStoryEstimatedCost(int managerId);
 
 	@Query("select min(us.estimatedCost) from UserStory us where us.manager.id = :managerId")
-	int minUserStoryEstimatedCost(int managerId);
+	double minUserStoryEstimatedCost(int managerId);
 
 	@Query("select max(us.estimatedCost) from UserStory us where us.manager.id = :managerId")
-	int maxUserStoryEstimatedCost(int managerId);
+	double maxUserStoryEstimatedCost(int managerId);
 
 	//Projects
